+++ mercurygui/__init__.py
@@ -1,4 +1,4 @@
 import mercurygui.config
 
-__version__ = "2.1.3"
+__version__ = "3.0.0.dev0"
 __author__ = "Sam Schott"