+++ mercurygui/main.py
@@ -16,16 +16,19 @@ import pkg_resources as pkgr
 import time
 import numpy as np
 import logging
+from pathlib import Path
 from PyQt5 import QtCore, QtWidgets, uic
 from mercuryitc.mercury_driver import MercuryITC_TEMP, MercuryITC_HTR, MercuryITC_AUX
 
 # local imports
-from mercurygui.feed import MercuryFeed
-from mercurygui.pyqt_labutils import LedIndicator, ConnectionDialog
-from mercurygui.pyqtplot_canvas import TemperatureHistoryPlot
-from mercurygui.config.main import CONF
+from .feed import MercuryFeed
+from .pyqt_labutils import LedIndicator, ConnectionDialog
+from .pyqtplot_canvas import TemperatureHistoryPlot
+from .config.main import CONF
 
 MAIN_UI_PATH = pkgr.resource_filename("mercurygui", "main.ui")
+PANEL_UI_PATH = pkgr.resource_filename("mercurygui", "panel.ui")
+MODULE_DIALOG_UI_PATH = pkgr.resource_filename("mercurygui", "module_dialog.ui")
 
 logger = logging.getLogger(__name__)
 
@@ -36,57 +39,26 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
     QUIT_ON_CLOSE = True
 
     MAX_DISPLAY = 24 * 60 * 60
+    UPDATE_FREQ = 1
     TITLE_TEMPLATE = "MercuryiTC Control"
+    log_path = Path.home() / ".mercurygui" / "LOG_FILES"
 
     def __init__(self, mercury):
         super(self.__class__, self).__init__()
         uic.loadUi(MAIN_UI_PATH, self)
 
         self.mercury = mercury
-        self.feed = MercuryFeed(mercury)
-
-        # sent Title font size relative to the system's default size
-        scaling = 1.5
-        font = self.labelTitle.font()
-        defaultFontSize = QtWidgets.QLabel("test").font().pointSize()
-        fontSize = round(defaultFontSize * scaling, 1)
-        font.setPointSize(fontSize)
-        self.labelTitle.setFont(font)
+        self._cached_connection_status = False
 
         # create popup Widgets
         self.connectionDialog = ConnectionDialog(self, self.mercury, CONF)
-        self.readingsDialog = None
-        self.modulesDialog = None
+        self.readingsDialog = ReadingsOverview(self.mercury)
+        self.modulesDialog = ModulesDialog(self.mercury)
 
         # create LED indicator
         self.led = LedIndicator(self)
         self.statusbar.addPermanentWidget(self.led)
         self.led.setChecked(False)
-
-        # set up temperature plot, adjust window margins accordingly
-        self.canvas = TemperatureHistoryPlot()
-        self.gridLayoutCanvas.addWidget(self.canvas)
-        w = self.canvas.y_axis_width
-        self.gridLayoutTop.setContentsMargins(w, 0, w, 0)
-        self.gridLayoutBottom.setContentsMargins(w, 0, w, 0)
-        self.horizontalSlider.setMaximum(self.MAX_DISPLAY / 60)
-
-        # connect slider to plot
-        self.horizontalSlider.valueChanged.connect(self.on_slider_changed)
-
-        # adapt text edit colors to graph colors
-        self.t1_reading.setStyleSheet("color:rgb%s" % str(self.canvas.GREEN))
-        self.gf1_edit.setStyleSheet("color:rgb%s" % str(self.canvas.BLUE))
-        self.h1_edit.setStyleSheet("color:rgb%s" % str(self.canvas.RED))
-        self.gf1_edit.setMinimalStep(0.1)
-        self.h1_edit.setMinimalStep(0.1)
-
-        # set up data vectors for plot
-        self.xdata = np.array([])
-        self.xdata_min_zero = np.array([])
-        self.ydata_tmpr = np.array([])
-        self.ydata_gflw = np.array([])
-        self.ydata_htr = np.array([])
 
         # restore previous window geometry
         self.restore_geometry()
@@ -96,54 +68,69 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
         self.showLogAction.triggered.connect(self.on_log_clicked)
         self.exitAction.triggered.connect(self.exit_)
         self.readingsAction.triggered.connect(self.on_readings_clicked)
-        self.connectAction.triggered.connect(self.feed.connect)
-        self.disconnectAction.triggered.connect(self.feed.disconnect)
         self.updateAddressAction.triggered.connect(self.connectionDialog.open)
+        self.connectionDialog.accepted.connect(self.build_tabs)
 
-        self.t2_edit.returnPressed.connect(self.change_t_setpoint)
-        self.r1_edit.returnPressed.connect(self.change_ramp)
-        self.r2_checkbox.clicked.connect(self.change_ramp_auto)
-        self.gf1_edit.returnPressed.connect(self.change_flow)
-        self.gf2_checkbox.clicked.connect(self.change_flow_auto)
-        self.gf3_edit.returnPressed.connect(self.change_flow_min)
-        self.h1_edit.returnPressed.connect(self.change_heater)
-        self.h2_checkbox.clicked.connect(self.change_heater_auto)
+        self.actionUpdateVeryOften.triggered.connect(lambda: self.set_update_freq(0.5))
+        self.actionUpdateOften.triggered.connect(lambda: self.set_update_freq(1))
+        self.actionUpdateNormally.triggered.connect(lambda: self.set_update_freq(2))
+
+        action_group = QtWidgets.QActionGroup(self)
+        action_group.addAction(self.actionUpdateVeryOften)
+        action_group.addAction(self.actionUpdateOften)
+        action_group.addAction(self.actionUpdateNormally)
 
         # initially disable menu bar items, will be enabled later individually
-        self.connectAction.setEnabled(True)
-        self.disconnectAction.setEnabled(False)
         self.modulesAction.setEnabled(False)
         self.readingsAction.setEnabled(False)
 
-        self.build_sensor_menu()
-
-        # initially disable controls, will be enabled later individually
-        self.t2_edit.setEnabled(False)
-        self.r1_edit.setEnabled(False)
-        self.r2_checkbox.setEnabled(False)
-        self.gf1_edit.setEnabled(False)
-        self.gf2_checkbox.setEnabled(False)
-        self.h1_edit.setEnabled(False)
-        self.h2_checkbox.setEnabled(False)
-
         # check if mercury is connected, connect slots
-        self.display_message(
-            "Looking for temperature controller at %s..." % self.mercury.visa_address
-        )
-        self.update_gui_connection(self.feed.connected)
+        self.display_message(f"Looking for Mercury at {self.mercury.visa_address}...")
 
-        # start (stop) updates of GUI when mercury is connected (disconnected)
-        # adjust clickable buttons upon connect / disconnect
-        self.feed.connected_signal.connect(self.update_gui_connection)
+        # populate panels for temperature modules
+        self.panels = {}
+        self.build_tabs()
 
-        # get new readings every second, update UI
-        self.feed.new_readings_signal.connect(self.update_controls)
-        self.feed.new_readings_signal.connect(self.update_plot)
+        self.update_timer = QtCore.QTimer()
+        self.update_timer.setInterval(self.UPDATE_FREQ * 1000)
+        self.update_timer.setSingleShot(False)  # set to reoccur
+        self.update_timer.timeout.connect(self.update_gui)
+        self.update_timer.start()
 
-        # set up logging to file
-        self.setup_logging()
+    def set_update_freq(self, seconds):
+        self.update_timer.setInterval(seconds * 1000)
 
-        self.feed.connect()
+        for panel in self.panels.values():
+            panel.feed.refresh = seconds
+
+    def build_tabs(self):
+
+        self.tabWidget.clear()
+
+        sensor_names = self._get_nicks(MercuryITC_TEMP)
+
+        if len(sensor_names) == 0:
+            sensor_names.append("Dummy")
+
+        for sensor_name in sensor_names:
+            panel = ControlPanel(self.mercury, self, sensor_name)
+            self.panels[sensor_name] = panel
+            self.tabWidget.addTab(panel, sensor_name)
+
+    def update_gui(self):
+
+        if not self.mercury.connected:
+            self.mercury.connect()
+
+        connected = self.mercury.connected
+
+        if connected is not self._cached_connection_status:
+            # update gui to reflect changed connection status
+            self.update_gui_connection(connected)
+            for panel in self.panels.values():
+                panel.update_gui_connection(connected)
+
+        self._cached_connection_status = self.mercury.connected
 
     # =================== BASIC UI SETUP ==========================================
 
@@ -163,7 +150,6 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
         CONF.set("Window", "y", geo.y())
 
     def exit_(self):
-        self.feed.exit_()
         self.save_geometry()
         self.deleteLater()
 
@@ -172,6 +158,148 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
             self.exit_()
         else:
             self.hide()
+
+    def update_gui_connection(self, connected):
+
+        if connected:
+            self.led.setChecked(True)
+
+            # enable / disable menu bar items
+            self.connectAction.setEnabled(False)
+            self.disconnectAction.setEnabled(True)
+            self.modulesAction.setEnabled(True)
+            self.readingsAction.setEnabled(True)
+            self.sensorAction.setEnabled(True)
+
+        elif not connected:
+            self.led.setChecked(False)
+
+            # enable / disable menu bar items
+            self.connectAction.setEnabled(True)
+            self.disconnectAction.setEnabled(False)
+            self.modulesAction.setEnabled(False)
+            self.readingsAction.setEnabled(False)
+
+    def display_message(self, text):
+        self.statusbar.showMessage(str(text), 5000)
+
+    def display_error(self, text):
+        self.statusbar.showMessage(str(text))
+
+    # ========================== CALLBACKS FOR MENU BAR ===========================
+
+    @QtCore.pyqtSlot()
+    def on_readings_clicked(self):
+        self.readingsDialog.update_gui()
+        self.readingsDialog.show()
+
+    @QtCore.pyqtSlot()
+    def on_module_selection_clicked(self):
+        self.modulesDialog.update_gui()
+        self.modulesDialog.open()
+
+    @QtCore.pyqtSlot()
+    def on_log_clicked(self):
+        """
+        Opens directory with log files with current log file selected.
+        """
+
+        if platform.system() == "Windows":
+            os.startfile(self.log_path)
+        elif platform.system() == "Darwin":
+            subprocess.Popen(["open", str(self.log_path)])
+        else:
+            subprocess.Popen(["xdg-open", str(self.log_path)])
+
+    def _get_nicks(self, sensor_type):
+        if self.mercury.connected:
+            return list(m.nick for m in self.mercury.modules if type(m) == sensor_type)
+        else:
+            return []
+
+
+# noinspection PyArgumentList
+class ControlPanel(QtWidgets.QMainWindow):
+
+    MAX_DISPLAY = 24 * 60 * 60
+
+    def __init__(self, mercury, parent, sensor_name=""):
+        super(self.__class__, self).__init__()
+        uic.loadUi(PANEL_UI_PATH, self)
+
+        self.mercury = mercury
+        self.parent = parent
+
+        self.sensor_name = sensor_name
+        self.temperature = self.get_modules(sensor_name)
+        if self.temperature:
+            self.feed = MercuryFeed(self.mercury, self.temperature, self.parent.UPDATE_FREQ)
+            self.feed.readings_signal.connect(self.update_gui)
+            self.feed.readings_signal.connect(self.update_plot)
+            self.feed.connected_signal.connect(self.update_gui_connection)
+
+        # set up temperature plot, adjust window margins accordingly
+        self.canvas = TemperatureHistoryPlot()
+        self.gridLayoutCanvas.addWidget(self.canvas)
+        self.horizontalSlider.setMaximum(self.MAX_DISPLAY / 60)
+
+        # connect slider to plot
+        self.horizontalSlider.valueChanged.connect(self.on_slider_changed)
+
+        # adapt text edit colors to graph colors
+        self.t1_reading.setStyleSheet("color:rgb%s" % str(self.canvas.GREEN))
+        self.gf1_edit.setStyleSheet("color:rgb%s" % str(self.canvas.BLUE))
+        self.h1_edit.setStyleSheet("color:rgb%s" % str(self.canvas.RED))
+        self.gf1_edit.setMinimalStep(0.1)
+        self.h1_edit.setMinimalStep(0.1)
+
+        # set up data vectors for plot
+        self.xdata = np.array([])
+        self.xdata_min_zero = np.array([])
+        self.ydata_tmpr = np.array([])
+        self.ydata_gflw = np.array([])
+        self.ydata_htr = np.array([])
+
+        # connect to callbacks
+        self.t2_edit.returnPressed.connect(self.change_t_setpoint)
+        self.r1_edit.returnPressed.connect(self.change_ramp)
+        self.r2_checkbox.clicked.connect(self.change_ramp_auto)
+        self.gf1_edit.returnPressed.connect(self.change_flow)
+        self.gf2_checkbox.clicked.connect(self.change_flow_auto)
+        self.gf3_edit.returnPressed.connect(self.change_flow_min)
+        self.h1_edit.returnPressed.connect(self.change_heater)
+        self.h2_checkbox.clicked.connect(self.change_heater_auto)
+
+        # enable or disable controls
+        self.update_gui_connection(self.mercury.connected)
+
+        # set up logging to file
+        self.setup_logging()
+
+    def get_modules(self, sensor_name):
+        """
+        Updates module list after the new modules have been selected.
+        """
+        # find all temperature modules
+        modules = getattr(self.mercury, "modules", [])
+        tmp_modules = [m for m in modules if type(m) is MercuryITC_TEMP]
+
+        # find match for given nick
+        match = next((m for m in tmp_modules if m.nick == sensor_name), None)
+
+        if match:
+            temperature = match
+        elif tmp_modules:
+            logger.warning(f'Sensor "{sensor_name}" not found, choosing first module')
+            temperature = tmp_modules[0]
+            self.sensor_name = tmp_modules[0].nick
+        else:
+            temperature = None
+            logger.warning('No temperature sensors found')
+
+        return temperature
+
+    # =================== BASIC UI SETUP ==========================================
 
     def on_slider_changed(self):
         # determine first plotted data point
@@ -182,34 +310,9 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
         self.canvas.p0.setXRange(-sv, 0)
         self.canvas.p0.enableAutoRange(x=False, y=True)
 
-    def build_sensor_menu(self):
-        self.sensorMenu = QtWidgets.QMenu()
-        self.sensorActionGroup = QtWidgets.QActionGroup(self)
-
-        for nick in self._get_nicks(MercuryITC_TEMP):
-            action = QtWidgets.QAction(nick)
-            action.setData(nick)
-            action.setCheckable(True)
-            action.setChecked(nick == self.feed.temperature_module_nick)
-            self.sensorMenu.addAction(action)
-            self.sensorActionGroup.addAction(action)
-            self.sensorActionGroup.triggered.connect(self.on_sensor_selected)
-
-        self.sensorAction.setMenu(self.sensorMenu)
-
-    @QtCore.pyqtSlot(bool)
     def update_gui_connection(self, connected):
 
         if connected:
-            self.display_message("Connection established.")
-            self.led.setChecked(True)
-
-            # enable / disable menu bar items
-            self.connectAction.setEnabled(False)
-            self.disconnectAction.setEnabled(True)
-            self.modulesAction.setEnabled(True)
-            self.readingsAction.setEnabled(True)
-            self.sensorAction.setEnabled(True)
 
             # enable controls
             self.t2_edit.setEnabled(True)
@@ -222,15 +325,6 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
             self.h2_checkbox.setEnabled(True)
 
         elif not connected:
-            self.display_error("Connection lost.")
-            logger.info("Connection to MercuryiTC lost.")
-            self.led.setChecked(False)
-
-            # enable / disable menu bar items
-            self.connectAction.setEnabled(True)
-            self.disconnectAction.setEnabled(False)
-            self.modulesAction.setEnabled(False)
-            self.readingsAction.setEnabled(False)
 
             # disable controls
             self.t2_edit.setEnabled(False)
@@ -242,23 +336,10 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
             self.h1_edit.setEnabled(False)
             self.h2_checkbox.setEnabled(False)
 
-        self.build_sensor_menu()
-
-    def display_message(self, text):
-        self.statusbar.showMessage("%s" % text, 5000)
-
-    def display_error(self, text):
-        self.statusbar.showMessage("%s" % text)
-
-    @QtCore.pyqtSlot(object)
-    def update_controls(self, readings):
+    def update_gui(self, readings):
         """
         Parses readings for the MercuryMonitorApp and updates UI accordingly
         """
-
-        self.labelTitle.setText(
-            self.TITLE_TEMPLATE + ": " + self.feed.temperature_module_nick
-        )
 
         # heater signals
         self.h1_label.setText("Heater, %s V:" % readings["HeaterVolt"])
@@ -311,7 +392,6 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
         else:
             self.alarm_label.hide()
 
-    @QtCore.pyqtSlot(object)
     def update_plot(self, readings):
         # append data for plotting
         self.xdata = np.append(self.xdata, time.time())
@@ -346,44 +426,41 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
             self.xdata, self.ydata_tmpr, self.ydata_gflw, self.ydata_htr
         )
 
+    def display_message(self, text):
+        self.parent.display_message(text)
+
+    def display_error(self, text):
+        self.parent.display_error(text)
+
     # =================== LOGGING DATA ============================================
 
     def setup_logging(self):
         """
-        Set up logging of temperature history to files.
-        Save temperature history to log file at '~/.CustomXepr/LOG_FILES/'
-        after every 10 min.
+        Save temperature history to log file at '~/.CustomXepr/LOG_FILES/' every 10 min.
         """
-        # find user home directory
-        home_path = os.path.expanduser("~")
-        self.logging_path = os.path.join(home_path, ".mercurygui", "LOG_FILES")
 
-        # create folder '~/.CustomXepr/LOG_FILES' if not present
-        if not os.path.exists(self.logging_path):
-            os.makedirs(self.logging_path)
+        os.makedirs(self.parent.log_path, exist_ok=True)
+
         # set logging file path
-        self.log_file = os.path.join(
-            self.logging_path,
-            "temperature_log " + time.strftime("%Y-%m-%d_%H-%M-%S") + ".txt",
-        )
+        time_str = time.strftime("%Y-%m-%d_%H-%M-%S")
+        self.log_file = self.parent.log_path / f"{self.sensor_name}_{time_str}.txt"
 
         # delete old log files
         now = time.time()
         days_to_keep = 7
 
-        for f in os.listdir(self.logging_path):
-            f = os.path.join(self.logging_path, f)
-            if os.stat(f).st_mtime < now - days_to_keep * 24 * 60 * 60:
-                if os.path.isfile(f):
-                    os.remove(f)
+        for file in os.scandir(self.parent.log_path):
+            if file.stat().st_mtime < now - days_to_keep * 24 * 60 * 60:
+                if file.is_file():
+                    os.remove(file.path)
 
         # set up periodic logging
         t_save = 10  # time interval to save temperature data (min)
-        self.save_timer = QtCore.QTimer()
-        self.save_timer.setInterval(t_save * 60 * 1000)
-        self.save_timer.setSingleShot(False)  # set to reoccur
-        self.save_timer.timeout.connect(self.log_temperature_data)
-        self.save_timer.start()
+        self.log_timer = QtCore.QTimer()
+        self.log_timer.setInterval(t_save * 60 * 1000)
+        self.log_timer.setSingleShot(False)  # set to reoccur
+        self.log_timer.timeout.connect(self.log_temperature_data)
+        self.log_timer.start()
 
     def save_temperature_data(self, path=None):
         # prompt user for file path if not given
@@ -391,11 +468,6 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
             text = "Select path for temperature data file:"
             path = QtWidgets.QFileDialog.getSaveFileName(caption=text)
             path = path[0]
-
-        if not path.endswith(".txt"):
-            path += ".txt"
-
-        title = "temperature trace, saved on " + time.strftime("%d/%m/%Y") + "\n"
 
         header = "\t".join(
             ["Time (sec)", "Temperature (K)", "Heater (%)", "Gas flow (%)"]
@@ -412,132 +484,73 @@ class MercuryMonitorApp(QtWidgets.QMainWindow):
         )
 
         # noinspection PyTypeChecker
-        np.savetxt(path, data_matrix, delimiter="\t", header=title + header, fmt="%f")
+        np.savetxt(path, data_matrix, delimiter="\t", header=header, fmt="%f")
 
     def log_temperature_data(self):
         # save temperature data to log file
-        if self.feed.connected:
+        if self.mercury.connected:
             self.save_temperature_data(self.log_file)
 
     # =================== CALLBACKS FOR SETTING CHANGES ===========================
 
-    @QtCore.pyqtSlot()
     def change_t_setpoint(self):
         new_t = self.t2_edit.value()
 
         if 3.5 < new_t < 300:
-            self.display_message("T_setpoint = %s K" % new_t)
-            self.feed.temperature.loop_tset = new_t
+            self.display_message(f"T_setpoint = {new_t} K")
+            self.temperature.loop_tset = new_t
         else:
             self.display_error(
-                "Error: Only temperature setpoints between "
-                + "3.5 K and 300 K allowed."
+                "Error: Only temperature setpoints between 3.5 K and 300 K allowed."
             )
 
-    @QtCore.pyqtSlot()
     def change_ramp(self):
-        self.feed.temperature.loop_rset = self.r1_edit.value()
-        self.display_message("Ramp = %s K/min" % self.r1_edit.value())
+        self.temperature.loop_rset = self.r1_edit.value()
+        self.display_message(f"Ramp = {self.r1_edit.value()} K/min")
 
-    @QtCore.pyqtSlot(bool)
     def change_ramp_auto(self, checked):
         if checked:
-            self.feed.temperature.loop_rena = "ON"
+            self.temperature.loop_rena = "ON"
             self.display_message("Ramp is turned ON")
         else:
-            self.feed.temperature.loop_rena = "OFF"
+            self.temperature.loop_rena = "OFF"
             self.display_message("Ramp is turned OFF")
 
-    @QtCore.pyqtSlot()
     def change_flow(self):
-        self.feed.temperature.loop_fset = self.gf1_edit.value()
-        self.display_message("Gas flow = %s%%" % self.gf1_edit.value())
+        self.temperature.loop_fset = self.gf1_edit.value()
+        self.display_message(f"Gas flow = {self.gf1_edit.value()}%%")
 
-    @QtCore.pyqtSlot()
     def change_flow_min(self):
-        self.feed.gasflow.gmin = self.gf3_edit.value()
-        self.display_message("Gas flow min = %s%%" % self.gf3_edit.value())
+        self.gasflow.gmin = self.gf3_edit.value()
+        self.display_message(f"Gas flow min = {self.gf3_edit.value()}%%")
 
-    @QtCore.pyqtSlot(bool)
     def change_flow_auto(self, checked):
         if checked:
-            self.feed.temperature.loop_faut = "ON"
+            self.temperature.loop_faut = "ON"
             self.display_message("Gas flow is automatically controlled.")
             self.gf1_edit.setReadOnly(True)
             self.gf1_edit.setEnabled(False)
         else:
-            self.feed.temperature.loop_faut = "OFF"
+            self.temperature.loop_faut = "OFF"
             self.display_message("Gas flow is manually controlled.")
             self.gf1_edit.setReadOnly(False)
             self.gf1_edit.setEnabled(True)
 
-    @QtCore.pyqtSlot()
     def change_heater(self):
-        self.feed.temperature.loop_hset = self.h1_edit.value()
-        self.display_message("Heater power  = %s%%" % self.h1_edit.value())
+        self.temperature.loop_hset = self.h1_edit.value()
+        self.display_message(f"Heater power  = {self.h1_edit.value()}%%")
 
-    @QtCore.pyqtSlot(bool)
     def change_heater_auto(self, checked):
         if checked:
-            self.feed.temperature.loop_enab = "ON"
+            self.temperature.loop_enab = "ON"
             self.display_message("Heater is automatically controlled.")
             self.h1_edit.setReadOnly(True)
             self.h1_edit.setEnabled(False)
         else:
-            self.feed.temperature.loop_enab = "OFF"
+            self.temperature.loop_enab = "OFF"
             self.display_message("Heater is manually controlled.")
             self.h1_edit.setReadOnly(False)
             self.h1_edit.setEnabled(True)
-
-    # ========================== CALLBACKS FOR MENU BAR ===========================
-
-    def on_sensor_selected(self, action):
-
-        nick = action.data()
-
-        try:
-            self.feed.select_temp_sensor(nick)
-        except IOError:
-            # the selected module name does not exists -> rebuild our list
-            self.build_sensor_menu()
-        else:
-            self.clear_plot()
-
-    @QtCore.pyqtSlot()
-    def on_readings_clicked(self):
-        # create readings overview window if not present
-        if self.readingsDialog is None:
-            self.readingsDialog = ReadingsOverview(self.mercury)
-        # show it
-        self.readingsDialog.show()
-
-    @QtCore.pyqtSlot()
-    def on_module_selection_clicked(self):
-        # create readings overview window if not present
-        if self.modulesDialog is None:
-            self.modulesDialog = ModulesDialog(self.mercury)
-        # show it
-        self.modulesDialog.update_gui()
-        self.modulesDialog.open()
-
-    @QtCore.pyqtSlot()
-    def on_log_clicked(self):
-        """
-        Opens directory with log files with current log file selected.
-        """
-
-        if platform.system() == "Windows":
-            os.startfile(self.logging_path)
-        elif platform.system() == "Darwin":
-            subprocess.Popen(["open", self.logging_path])
-        else:
-            subprocess.Popen(["xdg-open", self.logging_path])
-
-    def _get_nicks(self, sensor_type):
-        if self.feed.connected:
-            return list(m.nick for m in self.mercury.modules if type(m) == sensor_type)
-        else:
-            return []
 
 
 # noinspection PyUnresolvedReferences
@@ -605,14 +618,14 @@ class ReadingsOverview(QtWidgets.QDialog):
         super(self.__class__, self).__init__(parent=parent)
         self.setWindowTitle("Readings Overview")
         self.mercury = mercury
-        self.setupUi(self)
+        self.update_gui(self)
 
         # refresh readings every 3 sec
         self.timer = QtCore.QTimer()
         self.timer.timeout.connect(self.get_readings)
         self.timer.start(3000)
 
-    def setupUi(self, Form):
+    def update_gui(self, Form):
         Form.setObjectName("Mercury ITC Readings Overview")
         Form.resize(500, 142)
         self.masterGrid = QtWidgets.QGridLayout(Form)
@@ -625,7 +638,7 @@ class ReadingsOverview(QtWidgets.QDialog):
         # create a tab with combobox and text box for each module
         self.readings_tabs = []
 
-        for module in self.mercury.modules:
+        for module in getattr(self.mercury, "modules", []):
             new_tab = ReadingsTab(self.mercury, module)
             self.readings_tabs.append(new_tab)
             self.tabWidget.addTab(new_tab, module.nick)
@@ -651,19 +664,15 @@ class _NoModule:
 
 class ModulesDialog(QtWidgets.QDialog):
     """
-    Provides a user dialog to select the modules for the feed.
+    Provides a user dialog to select which gasflow and heater modules are associated
+    with a temperature sensor.
     """
 
     accepted = QtCore.pyqtSignal(object)
 
     def __init__(self, mercury, parent=None):
         super(self.__class__, self).__init__(parent=parent)
-        uic.loadUi(
-            os.path.join(
-                os.path.dirname(os.path.realpath(__file__)), "module_dialog.ui"
-            ),
-            self,
-        )
+        uic.loadUi(MODULE_DIALOG_UI_PATH, self)
 
         self.mercury = mercury
         self.update_gui()
@@ -677,26 +686,29 @@ class ModulesDialog(QtWidgets.QDialog):
         self.htr_modules.append(_NoModule())
         self.aux_modules.append(_NoModule())
 
+        self.comboBoxTEMP.clear()
+        self.comboBoxHTR.clear()
+        self.comboBoxAUX.clear()
+
         self.comboBoxTEMP.addItems([m.nick for m in self.temp_modules])
         self.comboBoxHTR.addItems([m.nick for m in self.htr_modules])
         self.comboBoxAUX.addItems([m.nick for m in self.aux_modules])
 
         # get current modules
-        self.comboBoxTEMP.setCurrentIndex(0)
-        self.comboBoxHTR.setCurrentText(self.temp_modules[0].loop_htr)
-        self.comboBoxAUX.setCurrentText(self.temp_modules[0].loop_aux)
+        if len(self.temp_modules) > 0:
+            self.comboBoxTEMP.setCurrentIndex(0)
+            self.comboBoxHTR.setCurrentText(self.temp_modules[0].loop_htr)
+            self.comboBoxAUX.setCurrentText(self.temp_modules[0].loop_aux)
 
         # connect callbacks
         self.comboBoxTEMP.currentIndexChanged.connect(self.on_temp_selected)
         self.buttonBox.accepted.connect(self.on_accept)
 
-    @QtCore.pyqtSlot(int)
     def on_temp_selected(self, index):
         # update content of heater and gasflow combo boxes
         self.comboBoxHTR.setCurrentText(self.temp_modules[index].loop_htr)
         self.comboBoxAUX.setCurrentText(self.temp_modules[index].loop_aux)
 
-    @QtCore.pyqtSlot()
     def on_accept(self):
         temp_index = self.comboBoxTEMP.currentIndex()
         htr_nick = self.comboBoxHTR.currentText()
@@ -715,8 +727,8 @@ class ModulesDialog(QtWidgets.QDialog):
         self.temp_modules[temp_index].loop_aux = aux_nick
 
     def _get_modules_for_type(self, sensor_type):
-
-        return [m for m in self.mercury.modules if type(m) is sensor_type]
+        modules = getattr(self.mercury, "modules", [])
+        return [m for m in modules if type(m) is sensor_type]
 
 
 def run():